Clear + Refactoring code styles
