+++ Core/Infrastructure/Models/MenuCommandItem.cs
@@ -1,11 +1,9 @@
 using System.Windows.Input;
-using ReactiveUI;
 
 namespace Core.Infrastructure.Models;
 
 public class MenuCommandItem
 {
-
     #region Fields
 
     private readonly Lazy<ICommand?> _lazyCommand;
@@ -19,12 +17,16 @@ public class MenuCommandItem
     public ICommand? Command => _lazyCommand.Value;
 
     #endregion
-    
+
+    #region Constructors
+
     public MenuCommandItem(string name, ICommand? command)
     {
         Name = name;
 
         _lazyCommand = new Lazy<ICommand?>(()=>command);
     }
-    
+
+
+    #endregion
 }
