+++ Core/Infrastructure/Services/NavigationService/BaseVmdNavigationService.cs
@@ -3,16 +3,14 @@ using Core.Infrastructure.Hosting;
 using Core.VMD.Base;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Core.Infrastructure.Services;
+namespace Core.Infrastructure.Services.NavigationService;
 
 public abstract class BaseVmdNavigationService<TBaseVMd> : BaseTypeNavigationService<TBaseVMd> where TBaseVMd : IBaseVmd
 {
-    public BaseVmdNavigationService(IStore<TBaseVMd> store) : base(store,getVmdFromService)
+    public BaseVmdNavigationService(IStore<TBaseVMd> store) 
+        : base(store, 
+            type =>(TBaseVMd)HostWorker.Services.GetRequiredService(type))
     {
     }
-
-    private static TBaseVMd getVmdFromService(Type vmdType) =>
-        (TBaseVMd)HostWorker.Services.GetRequiredService(vmdType);
-        
-
+    
 }
