+++ Core/Infrastructure/Services/NavigationService/TitleVmdNavigationService.cs
@@ -1,9 +1,9 @@
 using AppInfrastructure.Stores.DefaultStore;
 using Core.VMD.Base;
 
-namespace Core.Infrastructure.Services;
+namespace Core.Infrastructure.Services.NavigationService;
 
-public class TitleVmdsNavigationService : BaseVmdNavigationService<ITitleVmd>
+internal class TitleVmdsNavigationService : BaseVmdNavigationService<ITitleVmd>
 {
     public TitleVmdsNavigationService(IStore<ITitleVmd> store) : base(store)
     {
