+++ Core/Infrastructure/Stores/BaseVmdStore.cs
@@ -1,8 +1,9 @@
 using AppInfrastructure.Stores.DefaultStore;
+using Core.VMD.Base;
 
 namespace Core.Infrastructure.Stores;
 
-public abstract class BaseVmdStore<TBaseVmd> : BaseLazyStore<TBaseVmd>
+public abstract class BaseVmdStore<TBaseVmd> : BaseLazyStore<TBaseVmd> where TBaseVmd : IBaseVmd
 {
     
 }
