+++ Core/VMD/DevPanelVmds/DevVmd.cs
@@ -7,7 +7,7 @@ namespace Core.VMD.DevPanelVmds;
 
 public sealed class DevVmd : BaseVmd
 {
-    public LogsVmd? LogsPanelVmd { get; set; }
+    public LogsVmd? LogsPanelVmd { get; }
     
     public DevVmd()
     {
